% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{biogeographic_regions_eur}
\alias{biogeographic_regions_eur}
\title{Biogeographic regions of Europe}
\format{
A \code{data.frame} with 251138 observations and 4 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://zenodo.org/record/3766175#.XqbP5GgzZPZ}}
}
\usage{
data(biogeographic_regions_eur)
}
\description{
Biogeographical regions of Europe
}
\details{
A gird-based map for the biogeographical regions of Europe
}
\references{
Cervellini M, Zannini P, Di Musciano M, Fattorini S, Jimenez-Alfaro B, Rocchini D, Field R, R. Vetaas O,
Irl SD.H, Beierkuhnlein C, Hoffmann S, Fischer J-C, Casella L, Angelini P, Genovesi P, Nascimbene J, Chiarucci A (2020) 
A grid-based map for the Biogeographical Regions of Europe. Biodiversity Data Journal 8: e53720. 
\url{https://doi.org/10.3897/BDJ.8.e53720}
}
