% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{birdlife_eur_p11deg}
\alias{birdlife_eur_p11deg}
\title{Distribution of terrestrial birds at 0.11 degree spatial resolution for Europe}
\format{
A \code{data.frame} with 10728270 observations and 7 variables.
}
\description{
Data with x,y coordinates for each terrestrial bird species (present in 0.11 degree grid cells)
}
\details{
This dataset contains the distribution of all terrestrial bird species (on a 0.11 degree grid) for Europe
according to the IUCN range maps.
}
