% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{cci_eur}
\alias{cci_eur}
\title{Fraction of ESA CCI land cover time series (1992 - 2018) of Europe}
\format{
A \code{data.frame} with 249634 observations and 30 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://zenodo.org/record/3730469}}
}
\usage{
data(cci_eur)
}
\description{
Fraction of ESA CCI land cover time series (1992 - 2018) of Europe
}
\details{
Fraction of ESA CCI land cover time series (1992 - 2018) of Europe.
}
