% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{cordex_bioclim_eur_p44deg}
\alias{cordex_bioclim_eur_p44deg}
\title{Euro-cordex bioclimatic data of Europe}
\format{
A \code{data.frame} with 82782 observations and 27 variables.
}
\usage{
data(cordex_bioclim_eur_p44deg)
}
\description{
Euro-cordex bioclimatic data of Europe gridded to 0.44 degree
}
\details{
Euro-cordex climate simulation data of Europe.
The data.frame contains bioclimatic data for 3 time periods (1971-2000, 2021-2050, 2071-2100) 
under 3 representative concentration pathways (rcps; RCP2.6, RCP4.5, RCP8.5), 
5 global circulation models (gcms; MPI-M-MPI-ESM-LR, CNRM-CERFACS-CNRM-CM5, 
MOHC-HadGEM2-ES, ICHEC-EC-EARTH, IPSL-IPSL-CM5A-MR), 5 regional climate models (rcms; CLMcom-CCLM4-8-17, 
KNMI-RACMO22E, MPI-CSC-REMO2009, SMHI-RCA4, DMI-HIRHAM5), 
4 ensemble (r1i1p1, r2i1p1, r3i1p1, r12i1p1) and 3 rs (v1, v1a, v2).
}
