% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{corine_lc_eur_p44deg}
\alias{corine_lc_eur_p44deg}
\title{Corine land cover data of Europe}
\format{
A \code{data.frame} with 87563 observations and 8 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_lc_eur_p44deg)
}
\description{
Corine land cover data of Europe
}
\details{
Corine land cover data of Europe resampled onto a 0.44 degree grid.
Corine Land cover are land cover maps for Europe mapped at a resolution of 100 x 100 m.
}
