% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{e-obs_bioclim_1987_2017_eur}
\alias{e-obs_bioclim_1987_2017_eur}
\title{E-OBS gridded bioclimatic data for Europe at 0.1 degree resolution from 1987 to 2017}
\format{
A \code{data.frame} with 43008 observations and 21 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/insitu-gridded-observations-europe?tab=form}}
}
\usage{
data(`e-obs_bioclim_1987_2017_eur`)
}
\description{
E-OBS gridded bioclimatic data for Europe at 0.1 degree resolution from 1987 to 2017
}
\details{
E-OBS gridded bioclimatic data for Europe at 0.1 degree resolution from 1987 to 2017.
}
\references{
Cornes, R., G. van der Schrier, E.J.M. van den Besselaar, and P.D. Jones. 2018: 
An Ensemble Version of the E-OBS Temperature and Precipitation Datasets, 
J. Geophys. Res. Atmos., 123. doi:10.1029/2017JD028200
}
