% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{e-obs_elev_ens_eur_p1deg}
\alias{e-obs_elev_ens_eur_p1deg}
\title{E-OBS gridded elevation data for Europe at 0.1 degree resolution}
\format{
A \code{data.frame} with 44200 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/insitu-gridded-observations-europe?tab=form}}
}
\usage{
data(`e-obs_elev_ens_eur_p1deg`)
}
\description{
E-OBS gridded elevation data for Europe at 0.1 degree resolution
}
\details{
E-OBS gridded elevation data for Europe at 0.1 degree resolution.
}
\references{
Cornes, R., G. van der Schrier, E.J.M. van den Besselaar, and P.D. Jones. 2018: 
An Ensemble Version of the E-OBS Temperature and Precipitation Datasets, 
J. Geophys. Res. Atmos., 123. doi:10.1029/2017JD028200
}
