% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{ebba2_shp}
\alias{ebba2_shp}
\title{shapefile of EBBA2 grid}
\format{
A \code{sfc_MULTIPPOLYGON} object.
}
\usage{
data(ebba2_shp)
}
\description{
outline of EBBA2 grid
}
\details{
Boundaries of EBBA2 grid.
}
