% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{ebcc_eur}
\alias{ebcc_eur}
\title{European Bird Census Council data of Europe}
\format{
A \code{data.frame} with 3920 observations and 497 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://www.gbif.org}}. 
More information can be found here: \itemize{\item \url{https://www.ebcc.info/what-we-do/ebba2/ebcc-atlas-of-european-breeding-birds/}}.
}
\usage{
data(ebcc_eur)
}
\description{
EBCC bird data of Europe
}
\details{
European Breeding Bird Atlas data
}
