% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{ebcc_Sylvia_curruca}
\alias{ebcc_Sylvia_curruca}
\title{European Bird Census Council data of Sylvia curruca}
\format{
A \code{data.frame} with 2862 observations and 5 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://www.gbif.org}}. 
More information can be found here: \itemize{\item \url{https://www.ebcc.info/what-we-do/ebba2/ebcc-atlas-of-european-breeding-birds/}}.
}
\usage{
data(ebcc_Sylvia_curruca)
}
\description{
EBCC bird data of Sylvia curruca
}
\details{
European Breeding Bird Atlas data of Sylvia curruca
}
