% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{ebcc_Turdus_torquatus}
\alias{ebcc_Turdus_torquatus}
\title{European Bird Census Council data of Turdus torquatus}
\format{
A \code{data.frame} with 2642 observations and 5 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://www.gbif.org}}. 
More information can be found here: \itemize{\item \url{https://www.ebcc.info/what-we-do/ebba2/ebcc-atlas-of-european-breeding-birds/}}.
}
\usage{
data(ebcc_Turdus_torquatus)
}
\description{
EBCC bird data of Turdus torquatus
}
\details{
European Breeding Bird Atlas data of Turdus torquatus
}
