% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{package}
\name{edc package}
\alias{edc package}
\alias{edcpackage}
\title{R Package to access European species, climate and land-use data}
\description{
R Package to access European species, climate and land-use data
}
\references{
https://www.iucnredlist.org/resources/spatial-data-download; 
http://datazone.birdlife.org/species/spcdistPOS; 
https://datadryad.org/resource/doi:10.5061/dryad.83s7k
}
\author{
RS-eco
}
\keyword{package}
