% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{eu_dem_eur_p11deg}
\alias{eu_dem_eur_p11deg}
\title{Eurostat data of elevation across Europe}
\format{
A \code{data.frame} with 63258 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://ec.europa.eu/eurostat/de/web/gisco/geodata/reference-data/elevation/eu-dem/eu-dem-laea}}.
}
\usage{
data(eu_dem_eur_p11deg)
}
\description{
European elevation data at a resolution of 25 m resampled to 0.11 degree
}
\details{
European elevation data derived from Eurostat at a resolution of 25 m resampled to 0.11 degree.
}
