% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{eur_27}
\alias{eur_27}
\title{administrative boundary of EU27 countries}
\format{
A \code{sfc_MULTIPPOLYGON} object.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(eur_27)
}
\description{
GADM outline of EU27 countries
}
\details{
Administrative boundary of EU27 countries derived from GADM data.
}
