% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{eur_countries}
\alias{eur_countries}
\title{administrative boundary of EU27 countries plus Switzerland and UK}
\format{
A \code{sfc_MULTIPPOLYGON} object.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(eur_countries)
}
\description{
GADM outline of EU27 countries plus Switzerland and UK
}
\details{
Administrative boundary of EU27 countries plus Switzerland and UK 
derived from GADM data.
}
