% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{europe}
\alias{europe}
\title{administrative boundary of Europe}
\format{
A \code{sfc_MULTIPPOLYGON} object.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(europe)
}
\description{
GADM outline of Europe
}
\details{
Administrative boundary of Europe derived from GADM data.
}
