% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{hochkirch_orthoptera_eur}
\alias{hochkirch_orthoptera_eur}
\title{Distribution of Orthoptera species for Europe}
\format{
A \code{data.frame} with 1355 observations and 5 variables.
}
\description{
Data with range maps for each Orthoptera species in Europe
}
\details{
This dataset contains the distribution of all Orthoptera species in Europe
according to the range polygons by Hochkirch et al.
}
