% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{iucn_odonata_eur_p11deg}
\alias{iucn_odonata_eur_p11deg}
\title{Distribution of Odonata species at 0.11 degree spatial resolution for Europe}
\format{
A \code{data.frame} with 662221 observations and 7 variables.
}
\description{
Data with x,y coordinates for each Odonata species (present in 0.11 degree grid cells) in Europe
}
\details{
This dataset contains the distribution of all Odonata species (on a 0.11 degree grid) in Europe
according to the IUCN ranges.
}
