% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{kalkman_odonata_eur}
\alias{kalkman_odonata_eur}
\title{Distribution of Odonata species in Europe}
\format{
A \code{data.frame} with 6552 observations and 156 variables.
}
\description{
Data with x,y coordinates for each Odonata species in Europe
}
\details{
This dataset contains the distribution of Odonata species in Europe according to Kalkman et al.
}
