% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{landsystem_eur_1km}
\alias{landsystem_eur_1km}
\title{Landsystem data of Europe}
\format{
A \code{data.frame} with 4894729 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dataverse.nl/dataset.xhtml?persistentId=doi:10.34894/XNC5KA}}
}
\usage{
data(landsystem_eur_1km)
}
\description{
Landsystem data of Europe
}
\details{
Landsystem data of Europe
}
\references{
Dou, Y., Cosentino, F., Malek, Z. et al. (2021) 
A new European land systems representation accounting for landscape characteristics. Landscape Ecology. 
\url{https://doi.org/10.1007/s10980-021-01227-5}.
}
