% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{landsystem_perc_eur_10km}
\alias{landsystem_perc_eur_10km}
\title{Percentage cover landsystem data of Europe gridded to 10x10 km}
\format{
A \code{data.frame} with 57335 observations and 28 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dataverse.nl/dataset.xhtml?persistentId=doi:10.34894/XNC5KA}}
}
\usage{
data(landsystem_perc_eur_10km)
}
\description{
Percentage cover landsystem data of Europe gridded to 10x10 km
}
\details{
Percentage cover landsystem data of Europe gridded to 10x10 km
}
\references{
Dou, Y., Cosentino, F., Malek, Z. et al. (2021) 
A new European land systems representation accounting for landscape characteristics. Landscape Ecology. 
\url{https://doi.org/10.1007/s10980-021-01227-5}.
}
