% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{srtm_csi_eur_p44deg}
\alias{srtm_csi_eur_p44deg}
\title{srtm data of Europe}
\format{
A \code{data.frame} with 2996 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://srtm.csi.cgiar.org/srtmdata/}}
}
\usage{
data(srtm_csi_eur_p44deg)
}
\description{
NASA SRTM v3.0 elevation data of Europe
}
\details{
Elevation data of Europe derived from NASA SRTM 3 Arc Sec v3.0 data and aggregated to 0.44 degree resolution.
}
