% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edc-package.R
\docType{data}
\name{wfde5_snowf_eur}
\alias{wfde5_snowf_eur}
\title{WFDE5 reanalysis yearmon snowfall flux data of Europe}
\format{
A \code{data.frame} with 3507 observations and 494 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/10.24381/cds.20d54e34?tab=form}}
}
\usage{
data(wfde5_rainf_eur)
}
\description{
WFDE5 reanalysis yearmon snowfall flux data of Europe
}
\details{
WFDE5 reanalysis yearmon snowfall flux data of Europe.
The data.frame contains snowfall flux in kg m-2 s-1 from January 1979 - December 2019 at a 0.5° grid
}
